portal/web/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
colorlog==2.10.0
contextlib2==0.5.4        # via raven
django-environ==0.4.2
django==1.11              # via django-environ
httplib2==0.10.3          # via oauth2client
oauth2client==4.0.0
pyasn1-modules==0.0.8     # via oauth2client
pyasn1==0.2.3             # via oauth2client, pyasn1-modules, rsa
pytz==2017.2              # via django
raven==6.0.0
rsa==3.4.2                # via oauth2client
six==1.10.0               # via django-environ, oauth2client
django-jet
django-jfu==2.0.9
DjangoRestless
numpy==1.12
gunicorn==19.7.1
psycopg2==2.7.1
redis==2.10.5
Pillow==4.1.1
django-enumfields
django-polymorphic==1.2
django-model-render